requirements.txt
neo4j>=5.20.0
openai>=1.40.0
Pillow>=10.0.0
pydantic>=2.7.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
numpy>=1.26.0
streamlit>=1.37.0
plotly>=5.22.0
networkx>=3.3
fastapi>=0.114.0
uvicorn>=0.30.0
